Move Martingale class to a separate file
